+++ source/renderer/app/components/voting/voting-registration-wizard-steps/widgets/VotingRegistrationDialog.js
@@ -15,7 +15,7 @@ import type { DialogActions } from '../../../widgets/Dialog';
 const messages = defineMessages({
   dialogTitle: {
     id: 'voting.votingRegistration.dialog.dialogTitle',
-    defaultMessage: '!!!Register for Fund3 voting',
+    defaultMessage: '!!!Register for Fund4 voting',
     description: 'Tile "Register to vote" for voting registration',
   },
   subtitle: {